requirements.txt
h11~=0.14.0
pip~=24.3.1
attrs~=24.2.0
wheel~=0.42.0
PySocks~=1.7.1
protobuf~=4.25.2
cryptography~=42.0.2
pillow~=10.2.0
docutils~=0.21.2
filelock~=3.13.1
redis~=5.0.8
flet~=0.24.1
qrcode~=7.4.2
watchdog~=4.0.0
PyYAML~=6.0.1
packaging~=23.2
rich~=13.7.0
fastapi~=0.109.2
anyio~=4.2.0
starlette~=0.36.3
uvicorn~=0.27.1
pefile~=2023.2.7
pyinstaller~=6.10.0
Pygments~=2.17.2
markdown-it-py~=3.0.0
idna~=3.6
sniffio~=1.3.0
trio~=0.27.0
outcome~=1.3.0.post0
arrow~=1.3.0
python-dateutil~=2.8.2
typing_extensions~=4.12.2
click~=8.1.3
httpcore~=1.0.5
certifi~=2023.11.17
mdurl~=0.1.2
pydantic~=2.6.1
httpx~=0.27.0
MarkupSafe~=2.1.4
pypng~=0.20220715.0
lxml~=5.3.0
colorama~=0.4.6
gunicorn~=22.0.0
python-multipart~=0.0.9
chardet~=5.2.0
email_validator~=2.1.1
pydantic_core~=2.16.2
requests~=2.29.0
websockets~=11.0.3
python-slugify~=8.0.2
urllib3~=1.26.18
httptools~=0.6.1
wsproto~=1.2.0
watchfiles~=0.21.0
six~=1.16.0
oauthlib~=3.2.2
blinker~=1.7.0
annotated-types~=0.6.0
realtime~=1.0.2
charset-normalizer~=3.3.2
storage3~=0.7.0
supabase~=2.3.7
gotrue~=2.1.0
postgrest~=0.15.0
supafunc~=0.3.3
repath~=0.9.0
matplotlib~=3.8.2
deprecation~=2.1.0
itsdangerous~=2.1.2
Jinja2~=3.1.3
cookiecutter~=2.6.0
binaryornot~=0.4.4
python-dotenv~=1.0.1
pywin32~=308